refactor(ui): remove legacy button
