+++ packages/code-du-travail-css/src/components/_buttons.css
@@ -1,229 +1,3 @@
-/* Buttons.
-
-    <a href="#" class="btn">Valider</a>
-    <button type="submit" class="btn">Valider</button>
---------------------------------------------------------------------------- */
-
-.btn,
-.btn:link,
-.btn:visited {
-  appearance: none;
-
-  display: inline-block;
-  margin: 0;
-  text-align: center;
-
-  font-family: inherit;
-  font-size: inherit;
-  font-weight: bold;
-  line-height: inherit;
-  padding: var(--spacing-small) var(--spacing-base);
-  border-radius: var(--border-radius-base);
-  border: 1px solid transparent;
-
-  /* Fixme: currently use the same colors as `.btn__primary`. */
-  background: var(--color-blue-light);
-  border-bottom: 2px solid color-mod(var(--color-blue-light) lightness(-10%));
-  color: var(--color-primary-text);
-
-  cursor: pointer;
-
-  transition: background 250ms ease;
-
-  /* Specific to <a> elements. */
-  text-decoration: none;
-  &:active {
-    position: relative;
-    top: 1px;
-    left: 1px;
-  }
-
-  &:focus {
-    outline: none;
-  }
-  &:focus:not(:focus-visible) {
-    border: 1px solid var(--color-focus);
-  }
-  &:focus:not(:-moz-focusring) {
-    border: 1px solid var(--color-focus);
-  }
-
-  &[disabled] {
-    cursor: not-allowed;
-    opacity: 0.5;
-  }
-}
-
-.btn__primary {
-  background: var(--color-primary-background);
-  border-bottom: 2px solid
-    color-mod(var(--color-primary-background) lightness(-10%));
-  color: var(--color-primary-text);
-  &:focus,
-  &:hover,
-  &:active {
-    background-color: color-mod(
-      var(--color-primary-background) lightness(-10%)
-    );
-  }
-}
-
-.btn__secondary {
-  background-color: var(--color-secondary-background);
-  border-bottom: 2px solid
-    color-mod(var(--color-secondary-background) lightness(-10%));
-  color: var(--color-secondary-text);
-  &:focus,
-  &:hover,
-  &:active {
-    background-color: color-mod(
-      var(--color-secondary-background) lightness(-10%)
-    );
-  }
-}
-
-.btn__warning {
-  background-color: var(--color-warning-background);
-  border-bottom: 2px solid
-    color-mod(var(--color-warning-background) lightness(-10%));
-  color: var(--color-warning-text);
-  &:focus,
-  &:hover,
-  &:active {
-    background-color: color-mod(
-      var(--color-warning-background) lightness(-10%)
-    );
-  }
-}
-
-.btn__success {
-  background-color: var(--color-success-background);
-  border-bottom: 2px solid
-    color-mod(var(--color-success-background) lightness(-10%));
-  color: var(--color-success-text);
-  &:focus,
-  &:hover,
-  &:active {
-    background-color: color-mod(
-      var(--color-success-background) lightness(-10%)
-    );
-  }
-}
-
-.btn__info {
-  background-color: var(--color-info-background);
-  border-bottom: 2px solid
-    color-mod(var(--color-info-background) lightness(-10%));
-  color: var(--color-info-text);
-  &:focus,
-  &:hover,
-  &:active {
-    background-color: color-mod(var(--color-info-background) lightness(-10%));
-  }
-}
-
-.btn__danger {
-  background-color: var(--color-danger-background);
-  border-bottom: 2px solid
-    color-mod(var(--color-danger-background) lightness(-10%));
-  color: var(--color-danger-text);
-  &:focus,
-  &:hover,
-  &:active {
-    background-color: color-mod(var(--color-danger-background) lightness(-10%));
-  }
-}
-
-.btn-icon {
-  appearance: none;
-  border: none;
-  display: inline-block;
-  margin: 0;
-  text-align: center;
-  white-space: nowrap;
-
-  font-family: inherit;
-  font-size: inherit;
-  line-height: 0.8;
-  font-weight: normal;
-
-  padding: var(--spacing-xsmall) var(--spacing-small);
-  border-radius: var(--border-radius-base);
-  cursor: pointer;
-  transition: all 250ms ease;
-
-  background: transparent;
-  color: var(--color-dark-grey);
-
-  &:hover {
-    color: var(--color-blue-light);
-  }
-
-  &:active {
-    position: relative;
-    top: 1px;
-  }
-  &:focus {
-    color: var(--color-blue-dark);
-    background: var(--color-element-background);
-  }
-  &:focus:not(:focus-visible) {
-    background: transparent;
-    box-shadow: none;
-  }
-  &:focus:not(:-moz-focusring) {
-    background: transparent;
-    box-shadow: none;
-  }
-  &:focus-visible {
-    color: var(--color-black);
-    background: var(--color-element-background);
-    box-shadow: none;
-  }
-  &:-moz-focusring {
-    color: var(--color-black);
-    background: var(--color-element-background);
-    box-shadow: none;
-  }
-  & svg {
-    width: 30px;
-    height: 38px;
-  }
-}
-
-/* Buttons containing only 1 image.
-
-In this case, a text explaining what the button is doing should be made
-available to screen reader users, but hidden from sighted users, e.g.:
-
-    <button type="submit" class="btn btn__img btn__img__search">
-        <span class="hidden">Rechercher</span>
-    </button>
-
-The <span> element will be moved away from the normal flow of the document.
---------------------------------------------------------------------------- */
-
-.btn__img {
-  background-repeat: no-repeat;
-  background-position: 50% 50%;
-
-  /*
-    Artificially inject an empty text node with a width, otherwise the <button>
-    would be empty (because it's only child has a `.hidden` class) and properties
-    related to text (such as font-size, line-height, etc.) would not be applied.
-    This would result in buttons with incoherent sizes.
-    */
-  &::after {
-    content: " ";
-    min-width: 1rem;
-    display: inline-block;
-  }
-}
-.btn__img__search {
-  background-image: url(src/assets/icons/search.svg);
-  background-position: 50% 65%;
-}
-
 .btn-large {
   text-decoration: none;
   border: 1px var(--color-element-border) solid;
@@ -275,36 +49,4 @@ The <span> element will be moved away from the normal flow of the document.
   font-size: var(--font-size-xsmall);
   padding: var(--spacing-xsmall) var(--spacing-small);
   font-weight: 400;
-}
-
-.btn__link {
-  text-align: inherit;
-  text-decoration: underline;
-  padding: 0;
-  border: none;
-  border-radius: 0;
-  font-weight: inherit;
-  /* Fixme: currently use the same colors as `.btn__primary`. */
-  background: none;
-  color: currentColor;
-}
-
-.btn__link:focus,
-.btn__link:hover,
-.btn__link:active {
-  background: none;
-  text-decoration: none;
-  color: currentColor;
-}
-
-.link {
-  text-decoration: underline;
-  cursor: pointer;
-  color: var(--color-blue);
-}
-
-.link:focus,
-.link:hover,
-.link:active {
-  text-decoration: none;
 }
