+++ packages/code-du-travail-frontend/src/common/CompanyForm.js
@@ -62,10 +62,10 @@ const Company = ({ name, idccList, onClear }) => {
       <div>
         <b>{name}</b>
         <div>
-          <Button link onClick={onClear}>
+          <Button variant="link" onClick={onClear}>
             Chercher une autre entreprise
           </Button>
-          <Button link onClick={onClear}>
+          <Button variant="link" onClick={onClear}>
             {/* it does not seem to work when in the same Button */}
             <X size="16" />
           </Button>
