+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/index.js
@@ -98,7 +98,7 @@ class CalculateurIndemnite extends React.Component {
                   onClick={() => {
                     this.resetState(restart);
                   }}
-                  primary
+                  variant="primary"
                 >
                   recommencer
                 </Button>
